pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=44.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "NFACT"
version = "1.0.0"
description = "NFACT"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["NFACT"]

dependencies = [
    'scipy',
    'numpy',
    'progressbar',
    'pandas',
    'nibabel',
    'scikit-learn',
    'fslpy'

]
requires-python = ">=3.8"